laboratories/laboratory_4/1001326848/api_gateway/requirements.txt
flask
pyjwt
requests
